Correcting picard use after testing
Picard has been tested now and no longer deadlocks.
It is worth noting that picard produces some text as output of the program as well as generating metric files.  I am reading and outputting this to standard output right now but I might output to a log file later instead so that potential errors can be recorded (picard likes the input files to be organized in ways that the assemblers might not perform).